requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
executing==2.2.1
faiss-cpu==1.12.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.78
langchain-openai==0.3.35
langchain-tavily==0.2.12
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.32
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.3
openai==2.2.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==23.2
pandas==2.3.3
parso==0.8.5
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
pure_eval==0.2.3
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
